requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.3
pyTelegramBotAPI==3.7.7
python-decouple==3.4
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
